# historico_de_dados_da_aula/48 - http_fiber_api_gorm_orm-jwt-swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Este é um app construido junto com os alunos do torne-se um programador, com objetivo em aprender a linguagem Golang e seus componentes",
        "title": "Desafio de Golang",
        "contact": {
            "name": "Danilo Aparecido",
            "url": "https://www.torneseumprogramador.com.br/cursos/desafio_go_lang",
            "email": "suporte@torneseumprogramador.com.br"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "Mostrar dados sobre a API quando acessar a rota /",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Home"
                ],
                "summary": "Apresentação da API",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelviews.HomeView"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Autentica um administrador e retorna um token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "Login do Administrador",
                "parameters": [
                    {
                        "description": "Informações de Login",
                        "name": "loginDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.LoginDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token de autenticação",
                        "schema": {
                            "$ref": "#/definitions/modelviews.AdmToken"
                        }
                    },
                    "400": {
                        "description": "Erro ao analisar o corpo da requisição",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/recursos": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retorna uma lista de todos os recursos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recursos"
                ],
                "summary": "Lista todos os recursos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Recurso"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Cria um novo recurso com as informações fornecidas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recursos"
                ],
                "summary": "Cria um novo recurso",
                "parameters": [
                    {
                        "description": "Informações do Recurso",
                        "name": "recurso",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.RecursoDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Recurso"
                        }
                    }
                }
            }
        },
        "/recursos/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retorna detalhes de um recurso específico pelo ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recursos"
                ],
                "summary": "Mostra um recurso",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Recurso",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Recurso"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Atualiza o recurso com o ID fornecido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recursos"
                ],
                "summary": "Atualiza um recurso",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Recurso",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Informações atualizadas do Recurso",
                        "name": "recurso",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.RecursoDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Recurso"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Exclui o recurso com o ID fornecido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recursos"
                ],
                "summary": "Exclui um recurso",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Recurso",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
        "dtos.LoginDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "senha": {
                    "type": "string"
                }
            }
        },
        "dtos.RecursoDTO": {
            "type": "object",
            "properties": {
                "nome": {
                    "type": "string"
                },
                "valor": {
                    "type": "number"
                }
            }
        },
        "models.Recurso": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "valor": {
                    "type": "number"
                }
            }
        },
        "modelviews.AdmToken": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "modelviews.HomeView": {
            "type": "object",
            "properties": {
                "doc": {
                    "type": "string"
                },
                "mensagem": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}